Renamed AddActivityScreen component and correct the import statement in App.js
