+++ App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import BottomTabNavigator from './Components/TabNavigator';
-import AddAnActivityScreen from './Screens/AddAnActivityScreen';
+import AddActivityScreen from './Screens/AddActivityScreen';
 import AddDietScreen from './Screens/AddDietScreen';
 import EditActivityScreen from './Screens/EditActivityScreen';
 import EditDietScreen from './Screens/EditDietScreen';
@@ -31,7 +31,7 @@ export default function App() {
           />
           <Stack.Screen 
             name="Add An Activity" 
-            component={AddAnActivityScreen}
+            component={AddActivityScreen}
             options={{
               headerBackTitleVisible: false,
             }}
